Assimilate `anvil-plugin-babel` into `core`
