+++ examples/ft-app-jsx-express-style/anvil.config.json
@@ -1,6 +1,5 @@
 {
   "plugins": [
-    "@financial-times/anvil-plugin-babel",
     "@financial-times/anvil-plugin-code-splitting",
     "@financial-times/anvil-plugin-css",
     "@financial-times/anvil-plugin-ft-css",
