--- packages/anvil-plugin-babel/.npmignore
@@ -1,1 +1,0 @@
-__test__/
