--- packages/anvil-plugin-babel/.npmrc
@@ -1,2 +1,0 @@
-package-lock=false
-shrinkwrap=false
