--- packages/anvil-plugin-babel/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@financial-times/anvil-plugin-babel",
-  "version": "0.0.0",
-  "description": "",
-  "main": "dist/cjs/index.js",
-  "types": "src/index.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "tsc": "../../node_modules/.bin/tsc",
-    "clean": "npm run clean:dist && npm run clean:node_modules",
-    "clean:dist": "rm -rf dist",
-    "clean:node_modules": "rm -rf node_modules",
-    "build:cjs": "npm run tsc -- --outDir ./dist/cjs",
-    "build": "npm run clean:dist && npm run build:cjs",
-    "dev": "npm run clean:dist && npm run build:cjs -- --watch"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "peerDependencies": {
-    "@financial-times/anvil": "file:../anvil"
-  },
-  "dependencies": {
-    "@babel/core": "^7.1.2",
-    "@babel/preset-env": "^7.1.0",
-    "@financial-times/anvil-types-build": "file:../anvil-types-build",
-    "adonai": "v1.0.0-beta.5",
-    "babel-loader": "^8.0.4",
-    "webpack-merge": "^4.1.4"
-  },
-  "devDependencies": {
-    "@financial-times/anvil": "file:../anvil"
-  }
-}
