--- packages/anvil-plugin-babel/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "rootDir": "./src/",
-  "extends": "../../tsconfig.json"
-}
