+++ packages/anvil-plugin-ft-js/README.md
@@ -3,7 +3,7 @@
 ## Installation
 
 ```
-npm install --save-dev @financial-times/anvil-plugin-ft-js @financial-times/anvil-plugin-babel
+npm install --save-dev @financial-times/anvil-plugin-ft-js
 ```
 
 ## Usage
@@ -13,12 +13,10 @@ anvil.config.json
 ```json
 {
   "plugins": [
-    "@financial-times/anvil-plugin-babel", "@financial-times/anvil-plugin-ft-js"
+    "@financial-times/anvil-plugin-ft-js"
   ]
 }
 ```
-
-Note that `anvil-plugin-ft-js` extends `anvil-plugin-babel` and so it must be specified after `anvil-plugin-babel`
 
 ## Settings
